[FIX] auditlog: Change comparison with old field values.
Fix the comparison with old values when setting fields to empty values.
The old values for the comparison when using fast log mode were initialized
with False. This causes the problem that when a field is set to an emty value
by the user, auditlog does not create a log line because internal checks
compare False to False and do not consider logging the change. Full log mode
is not affected by this as there everything is logged by default. However,
even in fast log mode, some companies have regulatory requirements which
force them to log user actions that set empty values to fields.