[FIX] auditlog: Fix invalid reading of stale database values.
Due to the introduction of the ThrowAwayCache, in full log mode when attempting
to log a write, the reading of the 'new' values effectively reads the 'old'
values because prior changes were not flushed to the database yet at
that point. This resulted in missing log lines in the created log entry since
the old vs. new value comparison did not indicate any changes.